# services/workplaces/docs/openapi.yml
openapi: 3.0.1
info:
  x-audience: external-partner
  title: Workplace Management API
  description: API to manage workplaces, rooms and equipments
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://{stage}.domain.com
    description: Staged Domain Server
    variables:
      stage:
        description: Deployment Stage
        default: prod
        enum:
          - dev
          - qs
          - prod
tags:
  - name: workplaces
    description: Manage Workplaces
  - name: rooms
    description: Manage Rooms
  - name: equipments
    description: Manage Equipments
paths:
##############
# Equipments #
##############
  /equipments:
    get:
      operationId: getEquipments
      tags:
        - equipments
      summary: List available equipments
      responses:
        200:
          description: Equipments were read successfully
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/HALEquipmentsList'
  /equipments/{equipment}:
    parameters:
      - $ref: '#/components/parameters/equipmentName'
    get:
      operationId: getEquipment
      tags:
        - equipments
      summary: Get information about a single equipment by its name.
      responses:
        '200':
          description: Equipment was found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/HALEquipment'
        '404':
          description: Equipment does not exist
    put:
      operationId: replaceEquipment
      tags:
        - equipments
      summary: Create or update information about an equipment.
      requestBody:
        content:
          application/hal+json:
            schema:
              $ref: '#/components/schemas/Equipment'
        required: true
      responses:
        200:
          description: Equipment was updated successfully
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/HALEquipment'
        201:
          description: Equipment was created successfully
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/HALEquipment'
        400:
          description: Equipment is not valid (e.g. name does not match the expected pattern)
          content:
            application/problem+json:
              schema:
                $ref: '../../../docs/openapi/rfc-7807.yml#/components/schemas/Problem'
    delete:
      operationId: deleteEquipment
      tags:
        - equipments
      summary: Delete information about an equipment.
      responses:
        '204':
          description: Equipment was deleted successfully
        '404':
          description: Equipment does not exist
components:
  schemas:
    Equipment:
      type: object
      properties:
        name:
          type: string
          pattern: '^[a-z][-?[a-z0-9]+]*$'
          readOnly: true
        description:
          type: string
    HALEquipment:
      allOf:
        - $ref: '../../../docs/openapi/hal.yml#/components/schemas/HAL'
        - $ref: '#/components/schemas/Equipment'
    HALEquipmentsList:
      allOf:
        - $ref: '../../../docs/openapi/hal.yml#/components/schemas/HAL'
        - type: object
          properties:
            _embedded:
              type: object
              properties:
                equipments:
                  type: array
                  items:
                    $ref: '#/components/schemas/HALEquipment'
  parameters:
    equipmentName:
      name: equipment
      in: path
      description: The name of the equipment
      required: true
      schema:
        type: string
        pattern: '^[a-z][-?[a-z0-9]+]*$'